Progress (#34)
* Shader fixups

* Parse PLM

* Expand ANM parser to support files containing only Harry keyframes

* Fix dithering, improve draw function responsibilities

* Implement optional PSX vertex jitter

* Simplify uniforms

* 3D scene pixelization effect

* Shader fixups, restrict window aspect ratio and size

* Simplify post-process effect handling, add stubbed options, don't runble when vibration is disabled

* Implement FXAA

* Fix uneven pixels in dithering and pixelize shaders

* Minor fixups

* Revert pixelize and dither shader pixel scaling fix

* Fix text gradient

* Implement render scale

* Remove pixelise shader

* Flush

* Prepare extra render targets, update ping pong texture class

* String fixups, simplify sprite scale modes

* Make kdt-tool more legible

* Fix ported DMS code

* Integrate screen luma fade properly

* Remove gradient step code for glyphs

* Fix sprite and glyph scaling system

* Fail project build if any shaders fail

* Shader struct organisation, prepare 2D primitive shader for proper orthographic projection

* Fix sprite scaling once and for all

* Add fallback for esoteric TIMs

* Set bg sprite scale modes, fix video, fix race condition when fetching window size in renderer, simplify 2D sprite methods

* Use float versions of standard functions

* Function call fixups

* Renderer restructuring

* Fix text alignment modes

* Map message parser revision progress

* Sync deltas

* Map message parser progress, misc. changes

+++ Source/Input/Action.cpp
@@ -199,8 +199,8 @@ namespace Silent::Input
         float activeDelaySec   = (_ticksActive > SEC_TO_TICK(initialDelaySec)) ? delaySec : initialDelaySec;
         int   activeDelayTicks = SEC_TO_TICK(activeDelaySec);
 
-        int delayTicks     = (int)floor(_ticksActive     / activeDelayTicks) * activeDelayTicks;
-        int prevDelayTicks = (int)floor(_prevTicksActive / activeDelayTicks) * activeDelayTicks;
+        int delayTicks     = (int)floorf(_ticksActive     / activeDelayTicks) * activeDelayTicks;
+        int prevDelayTicks = (int)floorf(_prevTicksActive / activeDelayTicks) * activeDelayTicks;
         return delayTicks > prevDelayTicks;
     }
